+++ src/physics/convection.jl
@@ -5,12 +5,12 @@ export BettsMillerConvection
 """The simplified Betts-Miller convection scheme from Frierson, 2007,
 https://doi.org/10.1175/JAS3935.1. This implements the qref-formulation
 in their paper. Fields and options are $(TYPEDFIELDS)"""
-@kwdef struct BettsMillerConvection{NF} <: AbstractConvection
+@parameterized @kwdef struct BettsMillerConvection{NF} <: AbstractConvection
     "[OPTION] Relaxation time for profile adjustment"
     time_scale::Second = Hour(4)
 
     "[OPTION] Relative humidity for reference profile [1]"
-    relative_humidity::NF = 0.7
+    @param relative_humidity::NF = 0.7
 end
 
 Adapt.@adapt_structure BettsMillerConvection
@@ -384,21 +384,21 @@ export ConvectiveHeating
 """Convective heating as defined by Lee and Kim, 2003, JAS
 implemented as convection parameterization. Fields are
 $(TYPEDFIELDS)"""
-@kwdef struct ConvectiveHeating{NF, VectorType} <: AbstractConvection
+@parameterized @kwdef struct ConvectiveHeating{NF, VectorType} <: AbstractConvection
     "[OPTION] Q_max heating strength as 1K/time_scale"
     time_scale::Second = Hour(12)
 
     "[OPTION] Pressure of maximum heating [hPa]"
-    p₀::NF = 525
+    @param p₀::NF = 525 (bounds=Positive,)
 
     "[OPTION] Vertical extent of heating [hPa]"
-    σₚ::NF = 200
+    @param σₚ::NF = 200 (bounds=Positive,)
 
     "[OPTION] Latitude of heating [˚N]"
-    θ₀::NF = 0
+    @param θ₀::NF = 0 (bounds=-90..90,)
 
     "[OPTION] Latitudinal width of heating [˚]"
-    σθ::NF = 20
+    @param σθ::NF = 20 (bounds=Positive,)
 
     "[DERIVED] Latitudinal mask"
     lat_mask::VectorType