+++ src/physics/land/land.jl
@@ -28,15 +28,15 @@ end
 
 # LandModel defined through its components
 export LandModel
-@kwdef mutable struct LandModel{G, TD, T, SM, SN, V, R} <: AbstractWetLand
+@parameterized @kwdef mutable struct LandModel{G, TD, T, SM, SN, V, R} <: AbstractWetLand
     spectral_grid::SpectralGrid
-    geometry::G = LandGeometry(spectral_grid)
-    thermodynamics::TD = LandThermodynamics(spectral_grid)
-    temperature::T = LandBucketTemperature(spectral_grid)
-    soil_moisture::SM = LandBucketMoisture(spectral_grid)
-    snow::SN = SnowModel(spectral_grid)
-    vegetation::V = VegetationClimatology(spectral_grid)
-    rivers::R = nothing
+    @component geometry::G = LandGeometry(spectral_grid)
+    @component thermodynamics::TD = LandThermodynamics(spectral_grid)
+    @component temperature::T = LandBucketTemperature(spectral_grid)
+    @component soil_moisture::SM = LandBucketMoisture(spectral_grid)
+    @component snow::SN = SnowModel(spectral_grid)
+    @component vegetation::V = VegetationClimatology(spectral_grid)
+    @component rivers::R = nothing
 end
 
 # also allow spectral grid to be passed on as first an only positional argument to model constructors
@@ -66,11 +66,11 @@ variables(land::LandModel) = (
 )
 
 export DryLandModel
-@kwdef struct DryLandModel{G, TD, T} <: AbstractDryLand
+@parameterized @kwdef struct DryLandModel{G, TD, T} <: AbstractDryLand
     spectral_grid::SpectralGrid
-    geometry::G = LandGeometry(spectral_grid)
-    thermodynamics::TD = LandThermodynamics(spectral_grid)
-    temperature::T = LandBucketTemperature(spectral_grid)
+    @component geometry::G = LandGeometry(spectral_grid)
+    @component thermodynamics::TD = LandThermodynamics(spectral_grid)
+    @component temperature::T = LandBucketTemperature(spectral_grid)
 end
 
 function initialize!(land::DryLandModel, model::PrimitiveEquation)