+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Changed OneBandShortwave to ij logic [#955](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/955)
+- Follow-up to #783: finalize @param/@component markers [#935](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/935)
 - Adjust fill!(::PrognosticVariables, x) to new variable system [#949](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/949)
 - Code formatting with Runic applied [#950](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/950)
 - Implicit correction kernelized and GPU-ready [#891](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/891)